move table to semanticcpg and make it a case class
